refactor: Change the location of the formCreateUser store to be centralized in a folder
