+++ src/components/user/form/UserForm.tsx
@@ -4,7 +4,7 @@ import EditUserForm from '@/components/user/form/EditUserForm';
 import SkeletonFormCreateUser from '@/components/user/form/skeletons/SkeletonFormCreateUser';
 import { useGetUserForEdit } from '@/hooks/useGetUserForEdit';
 import { useFetchCatalogInitCreateUserQuery } from '@/lib/queries/user';
-import { useFormCreateUserStore } from '@/lib/store/formCreateUser';
+import { useFormCreateUserStore } from '@/lib/store/forms/formCreateUser';
 import { Button, Result } from 'antd';
 import { useSearchParams } from 'next/navigation';
 import { useEffect, useMemo } from 'react';
