+++ src/lib/queries/role.ts
@@ -3,7 +3,7 @@ import { keepPreviousData, useQuery } from '@tanstack/react-query';
 
 export const useFetchListRoleQuery = () => {
   return useQuery({
-    queryKey: ['list-catalog-role'],
+    queryKey: ['list-roles'],
     queryFn: () => ListRolesService(),
     placeholderData: keepPreviousData,
   });
