+++ src/styles/modules/pecs.module.scss
@@ -1,0 +1,16 @@
+.phase_row_card_mobile {
+  @extend .row-between-start;
+  border: $grey-dark solid 1px;
+  padding: 8px 12px;  
+  border-radius: 10px;  
+
+  .text_name{
+    color: $primary-color-text;
+    font-size: 16px;
+    font-weight: bold;
+  }
+  .text_description{
+    color: $grey-dark;
+    font-size: 14px;
+  }
+}
